+++ priv/repo/migrations/20200712210146_add_attachment_filename_setting_to_users.exs
@@ -1,0 +1,9 @@
+defmodule Pleroma.Repo.Migrations.AddAttachmentFilenameSettingToUsers do
+  use Ecto.Migration
+
+  def change do
+    alter table(:users) do
+      add(:show_attachment_filenames, :boolean, default: false)
+    end
+  end
+end