+++ test/pleroma/web/mastodon_api/update_credentials_test.exs
@@ -180,6 +180,16 @@ defmodule Pleroma.Web.MastodonAPI.UpdateCredentialsTest do
       assert refresh_record(user).skip_thread_containment
     end
 
+    test "updates the user's show_attachment_filenames option", %{user: user, conn: conn} do
+      response =
+        conn
+        |> patch("/api/v1/accounts/update_credentials", %{show_attachment_filenames: "true"})
+        |> json_response_and_validate_schema(200)
+
+      assert response["pleroma"]["show_attachment_filenames"]
+      assert refresh_record(user).show_attachment_filenames
+    end
+
     test "updates the user's hide_follows status", %{conn: conn} do
       conn = patch(conn, "/api/v1/accounts/update_credentials", %{hide_follows: "true"})