+++ CHANGELOG.md
@@ -273,6 +273,7 @@ switched to a new configuration mechanism, however it was not officially removed
 - Mastodon API: Support irreversible property for filters.
 - Mastodon API: Add pleroma.favicon field to accounts.
 - Mastodon API: Add `filename` parameter to `POST /api/v1/media` and `POST /api/v2/media`.
+- Mastodon API: Add account setting for providing filename to post attachments
 - Admin API: endpoints for create/update/delete OAuth Apps.
 - Admin API: endpoint for status view.
 - OTP: Add command to reload emoji packs